requirements.txt
annotated-types==0.7.0
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
execnet==2.1.1
Faker==37.3.0
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
pluggy==1.5.0
psycopg2==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
pytest==8.3.5
pytest-html==4.1.1
pytest-metadata==3.1.1
pytest-rerunfailures==16.1
pytest-xdist==3.6.1
python-dotenv==1.1.0
requests==2.32.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.4.0
